strict load = False, move ckpt para match to convert_pllava

+++ examples/opensora_hpcai/tools/caption/PLLaVA/models/pllava/configuration_pllava.py
@@ -23,7 +23,7 @@ class PllavaConfig:
         use_pooling=True,
         gradient_checkpointing=False,
         pad_token_id=0,
-        dtype=ms.bfloat16,
+        dtype=ms.float32,
         **kwargs,
     ):
         self.ignore_index = ignore_index
